requirements.txt
py-clob-client>=0.25.0
httpx>=0.27.0
pydantic>=2.7.0
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
rich>=13.7.0
sqlalchemy>=2.0.29
python-dateutil>=2.9.0
